Simple_Program_to Convert PDF to AudioBooks.
Learning a Research Paper or E-books is one of the Least Attracting Things we do. Though, we hate it, we definitely using it daily. Thus, being a Python Developer, let's convert it to a Free to use Audio-Book. Let's only use our sharp ears rather than scrolling our eyeballs through a magnifier. Also, will be helpful to Bling guy who wants to learn a 25 paged AI Research papers and more.